Test the main loop kernels
Take advantage of some recent fixes in the main branch, and implement tests for
the 3 main loop thread kernels.

The intended new tests are in `test_interchange_with_rabbit.py`, but to support
that, add `utils.py` and rearrange the conftest fixtures.
