+++ funcx_endpoint/tests/funcx_endpoint/endpoint/test_interchange_with_rabbit.py
@@ -1,154 +1,144 @@
-import logging
-import multiprocessing
-import os
-import pickle
-import time
+from __future__ import annotations
 
+import multiprocessing
+import time
+import uuid
+import warnings
+from unittest import mock
+
+import dill
 import pika
 import pytest
-from parsl.providers import LocalProvider
-from tests.test_rabbit_mq.result_queue_subscriber import ResultQueueSubscriber
-from tests.test_rabbit_mq.task_queue_publisher import TaskQueuePublisher
+from funcx_common.messagepack import pack
+from funcx_common.messagepack.message_types import Result, Task
+from tests.funcx_endpoint.executors.mock_executors import MockExecutor
+from tests.utils import try_for_timeout
 
 from funcx_endpoint.endpoint.interchange import EndpointInterchange
 from funcx_endpoint.endpoint.register_endpoint import register_endpoint
-from funcx_endpoint.endpoint.results_ack import ResultsAckHandler
 from funcx_endpoint.endpoint.utils.config import Config
-from funcx_endpoint.executors import HighThroughputExecutor
-from funcx_endpoint.executors.high_throughput.messages import EPStatusReport, Task
-
-endpoint_dir = f"{os.getcwd()}/test_interchange_e2e"
-
-config = Config(
-    executors=[
-        HighThroughputExecutor(
-            max_workers_per_node=1,
-            provider=LocalProvider(
-                init_blocks=1,
-                min_blocks=0,
-                max_blocks=1,
-            ),
-            run_dir=endpoint_dir,
-        )
-    ],
-    funcx_service_address="https://api2.funcx.org/v2",
-)
 
 
-def make_mock_task():
-    task_buf = (
-        b"\x05TID=16a3b7d0-1c7f-4701-b3d5-dc22e8819f79;CID=RAW;"
-        b"85\n04\ngANYBgAAAGRvdWJsZXEAWCAAAABkZWYgZG91YmxlKHgpO"
-        b"gogICAgcmV0dXJuIHggKiAyCnEBhnEC\nLg==\n16\n00\ngANLAI"
-        b"VxAC4=\n12\n00\ngAN9cQAu\n"
-    )
-    task = Task(
-        task_id="16a3b7d0-1c7f-4701-b3d5-dc22e8819f79",
-        container_id="RAW",
-        task_buffer=task_buf,
-    )
-    return task
-
-
-def run_ix_process(reg_info, endpoint_uuid):
-    results_ack_handler = ResultsAckHandler(
-        endpoint_dir=endpoint_dir,
-    )
-    logging.warning("Starting Interchange process")
-    ix = EndpointInterchange(
-        config=config,
-        endpoint_id=endpoint_uuid,
-        results_ack_handler=results_ack_handler,
-        reg_info=reg_info,
-        endpoint_dir=endpoint_dir,
-    )
-
-    logging.warning("IX created, starting")
-    ix.start()
-    logging.warning("Interchange exiting")
-
-
-@pytest.mark.skip()
-def test_endpoint_interchange_against_rabbitmq(
-    endpoint_uuid, get_standard_funcx_client, setup_register_endpoint_response
+@pytest.fixture
+def run_interchange_process(
+    get_standard_funcx_client, setup_register_endpoint_response, tmp_path
 ):
-    """This test registers the endpoint against local RabbitMQ, and
-    uses mock tasks
     """
-    logging.warning(f"Current proc pid: {os.getpid()}")
-    logging.warning(f"Using endpoint_uuid: {endpoint_uuid}")
+    Start and stop a subprocess that executes the EndpointInterchange class.
 
-    # Use register_endpoint to get connection params
+    Yields a tuple of the interchange subprocess, (temporary) working directory,
+    a random endpoint id, and the mocked registration info.
+    """
+
+    def run_it(reg_info, endpoint_uuid, endpoint_dir, mock_funcx_cli=True):
+        fake_client = mock.Mock() if mock_funcx_cli else None
+
+        mock_exe = MockExecutor()
+        mock_exe.endpoint_id = endpoint_uuid
+
+        config = Config(
+            executors=[mock_exe],
+            funcx_service_address="https://api2.funcx.org/v2",
+        )
+
+        base_ref_mock = "funcx_endpoint.endpoint.interchange"
+        with mock.patch(f"{base_ref_mock}.register_endpoint") as mock_reg:
+            mock_reg.return_value = reg_info
+            ix = EndpointInterchange(
+                config=config,
+                endpoint_id=endpoint_uuid,
+                funcx_client=fake_client,
+                endpoint_dir=endpoint_dir,
+            )
+
+            ix.start()
+
+    endpoint_uuid = str(uuid.uuid4())
     endpoint_name = "endpoint_foo"
     fxc = get_standard_funcx_client()
+    setup_register_endpoint_response(endpoint_uuid)
     reg_info = register_endpoint(
         fxc,
         endpoint_uuid=endpoint_uuid,
-        endpoint_dir=os.getcwd(),
+        endpoint_dir=tmp_path,
         endpoint_name=endpoint_name,
     )
-    assert isinstance(reg_info, tuple)
-    assert len(reg_info) == 2
-    for param in reg_info:
-        assert isinstance(param, dict)
-        assert "pika_conn_params" in param
-        assert isinstance(param["pika_conn_params"], pika.URLParameters)
+    assert isinstance(reg_info, tuple), "Test setup verification"
 
-    task_queue_info, result_queue_info = reg_info
-
-    # Start the service side components:
-    # Connect the TaskQueuePublisher and submit some mock tasks
-    task_q_out = TaskQueuePublisher(
-        endpoint_id=endpoint_uuid,
-        conn_params=task_queue_info["pika_conn_params"],
-    )
-    task_q_out.connect()
-    task_q_out._channel.queue_purge(task_q_out.queue_name)
-    logging.warning(f"Publishing task to {endpoint_uuid}")
-
-    result_q_in = multiprocessing.Queue()
-    result_sub_proc = ResultQueueSubscriber(
-        conn_params=result_queue_info["pika_conn_params"],
-        external_queue=result_q_in,
-    )
-    result_sub_proc.start()
-
-    # Make a Task and publish it
-    task = make_mock_task()
-    task_q_out.publish(task.pack())
-
-    # Create the endpoint
     ix_proc = multiprocessing.Process(
-        target=run_ix_process,
-        args=(
-            reg_info,
-            endpoint_uuid,
-        ),
+        target=run_it, args=(reg_info, endpoint_uuid), kwargs={"endpoint_dir": tmp_path}
     )
     ix_proc.start()
-    time.sleep(2)
 
-    for _i in range(10):
-        from_ep, b_message = result_q_in.get()
-        # logging.warning(f"Got message: {message} from {from_ep}")
+    yield ix_proc, tmp_path, endpoint_uuid, reg_info
 
-        assert from_ep == endpoint_uuid
-        try:
-            message = pickle.loads(b_message)
-        except Exception:
-            logging.warning(f"Unknown message: {b_message} from {from_ep}")
-            continue
+    if ix_proc.is_alive():
+        ix_proc.terminate()
+        try_for_timeout(lambda: not ix_proc.is_alive())
 
-        if isinstance(message, EPStatusReport):
-            continue
+    rc = ix_proc.exitcode
+    if rc is not None and rc != 0:
+        warnings.warn(f"Interchange process exited with nonzero result code: {rc}")
 
-        if "result" in message or "exception" in message:
-            logging.warning(f"Received result/exception for {message['task_id']}")
-            assert message["task_id"] == task.task_id
-            break
+    if rc is None:
+        warnings.warn("Interchange process did not shut down cleanly - send SIGKILL")
+        ix_proc.kill()
 
-    logging.warning("Exiting...")
-    result_sub_proc.terminate()
-    task_q_out.close()
+
+def test_epi_graceful_shutdown(run_interchange_process):
+    ix_proc, tmp_path, endpoint_uuid, _reg_info = run_interchange_process
+    time.sleep(2)  # simple test approach for now: assume it's up after 2s
     ix_proc.terminate()
-    ix_proc.kill()
+    assert try_for_timeout(lambda: ix_proc.exitcode is not None), "Failed to shutdown"
+
+
+def test_epi_stored_results_processed(run_interchange_process):
+    ix_proc, tmp_path, endpoint_uuid, _reg_info = run_interchange_process
+
+    unacked_results_dir = tmp_path / "unacked_results"
+    unacked_results_dir.mkdir(exist_ok=True)
+    stored_task_name = str(uuid.uuid4())
+    stored_task_path = unacked_results_dir / stored_task_name
+    stored_task_path.write_bytes(b"GIBBERISH")
+
+    def file_is_gone():
+        return not stored_task_path.exists()
+
+    assert try_for_timeout(file_is_gone), "Expected stored task to be handled"
+
+
+def test_epi_forwards_tasks_and_results(
+    run_interchange_process, pika_conn_params, randomstring
+):
+    """
+    Verify the two main threads of kernel interest: that tasks are pulled from the
+    appropriate queue, and results are put into appropriate queue with the correct
+    routing_key.
+    """
+    ix_proc, tmp_path, endpoint_uuid, reg_info = run_interchange_process
+
+    task_uuid = uuid.uuid4()
+    task_msg = Task(task_id=task_uuid, task_buffer=randomstring())
+
+    task_q, res_q = reg_info
+    res_q_name = res_q["queue"]
+    task_q_name = task_q["queue"]
+    task_exch = task_q["exchange"]
+    with pika.BlockingConnection(pika_conn_params) as mq_conn:
+        with mq_conn.channel() as chan:
+            chan.queue_purge(task_q_name)
+            chan.queue_purge(res_q_name)
+            # publish our canary task
+            chan.basic_publish(task_exch, task_q_name, pack(task_msg))
+
+            # then get (consume) our expected result
+            result: Result | None = None
+            for mframe, mprops, mbody in chan.consume(
+                queue=res_q_name, inactivity_timeout=5
+            ):
+                assert (mframe, mprops, mbody) != (None, None, None), "no timely result"
+                result = dill.loads(mbody)
+                break
+    assert result is not None
+    assert result.task_id == task_uuid
+    assert result.data == task_msg.task_buffer
