+++ funcx_endpoint/tests/unit/test_endpointinterchange.py
@@ -4,7 +4,7 @@ from funcx_endpoint.endpoint.interchange import EndpointInterchange
 
 
 def test_main_exception_always_quiesces(mocker, tmp_path):
-    false_true_g = iter((False, True))
+    false_true_g = iter((False, False, True))
 
     def false_true():
         return next(false_true_g)
@@ -17,7 +17,6 @@ def test_main_exception_always_quiesces(mocker, tmp_path):
         config=mock_conf,
         logdir=str(tmp_path),
         endpoint_dir=str(tmp_path),
-        results_ack_handler=MagicMock(),
     )
     ei._task_puller_proc = MagicMock()
     ei._start_threads_and_main = MagicMock()
